Initial attempt at Perlin noise
Did a little hack where I instantiate little prefabs cubes in front of the white cube. I should probably write some code for the shader to do its thing instead of this workaround, but for now this is what we have.

The result is kind of splotchy for some reason-- not sure if it's a bug or just a by-product of the version of Perlin I'm using (apparently there's an improved version). I'm following the one from wikipedia.

And the red sphere at the top left corner of the white box was used to help me find (0,0) in box coordinates-- I'm thinking of that like a canvas.